Refactor timeline code to use TimelineController and remove html dependency.
add new line.
