+++ lib/timeline/timeline_protocol.dart
@@ -117,6 +117,9 @@ class TimelineThread implements Comparable<TimelineThread> {
 
   TimelineThreadEvent durationStack;
 
+  bool get isDartThread => _name == 'io.flutter.1.ui';
+  bool get isGpuThread => _name == 'io.flutter.1.gpu';
+
   final Map<String, TimelineThreadEvent> _asyncEvents =
       <String, TimelineThreadEvent>{};
